Macro Argy:Chile/requirements.txt
# Core data + storage
pandas>=2.1
numpy>=1.26
pyarrow>=14.0
pyyaml>=6.0
python-dateutil>=2.8

# Bloomberg
xbbg>=0.7.7
blpapi>=3.20

# Haver Analytics — installs with DLX, not from pip.
# If `import Haver` fails, ensure DLX is installed and Haver's Python folder
# (typically C:\DLX\Data) is on PYTHONPATH.

# Excel output
openpyxl>=3.1
xlsxwriter>=3.1

# Dashboard
streamlit>=1.30
plotly>=5.18

# CLI / logging
click>=8.1
rich>=13.7
